scripts/pubsub_tool_status_dataflow_version/requirements.txt
apache_beam==2.47.0
elasticsearch==7.16.2
python-dotenv==1.0.0
google-cloud-pubsub==2.16.0
google-cloud-secret-manager==2.16.1